requirements.txt
fastapi == 0.135.3
langgraph == 1.1.6
pydantic == 2.12.5
google-genai == 1.72.0
langchain-google-genai == 4.2.1
langchain == 1.2.15
langchain_core == 1.2.28
langchain_community == 0.3.31
dotenv == 0.9.9
faiss-cpu == 1.13.2
pypdf == 6.10.0
pymupdf == 1.27.2.2
